Revert "Merge pull request #619 from davidhao3300/master"
This reverts commit ce4c3e96066df68b5d47e96ba1aeb80b9f61d9a5, reversing
changes made to 71fb40a1bcd1da36b37de5677a9423d143d70c98.
