--- docs/xunit.md
@@ -1,6 +1,0 @@
-Note on the XUnit Reporter
-==========================
-
-The XUnit reporter will write intermediate test results to stdout, except when
-stdout is being used to output the test summary XML. In this case, intermediate
-test results will not be printed out.
