+++ lib/ci/test_reporters/xunit_reporter.js
@@ -1,10 +1,9 @@
-var displayutils = require('../../displayutils')
+var strutils = require('../../strutils')
 var XmlDom = require('xmldom')
 
 
 function XUnitReporter(silent, out){
   this.out = out || process.stdout
-  this.intermediateOut = this.out == process.stdout ? null : process.stdout
   this.silent = silent
   this.stoppedOnError = null
   this.id = 1
@@ -20,7 +19,6 @@ XUnitReporter.prototype = {
       launcher: prefix,
       result: data
     })
-    this.display(prefix, data)
     this.total++
     if (data.passed) this.pass++
   },
@@ -46,10 +44,6 @@ XUnitReporter.prototype = {
     }
 
     return new XmlDom.XMLSerializer().serializeToString(doc.documentElement);
-  },
-  display: function(prefix, result){
-    if (this.silent || !this.intermediateOut) return
-    this.intermediateOut.write(displayutils.resultString(this.id++, prefix, result))
   },
   getTestResultNode: function(document, result){
     var launcher = result.launcher
